Crate file_url

The URL crate parses file:// URLs differently on Windows and other operating systems. This crates provides functionality that tries to parse a file:// URL as a path on all operating systems. This is useful when you want to convert a file:// URL to a path and vice versa.

Enums

FileURLParseError

Functions

directory_path_to_url

file_path_to_url

is_windows_drive_letter

Whether the scheme is file:, the path has a single segment, and that segment is a Windows drive letter

url_to_path

Tries to convert a file:// based URL to a path.

url_to_typed_path

Tries to convert a file:// based URL to a path.