file_url

Function url_to_typed_path

pub fn url_to_typed_path(url: &Url) -> Option<Utf8TypedPathBuf>

Tries to convert a file:// based URL to a path.

We assume that any passed URL that represents a path is an absolute path.

Url::to_file_path has a different code path for Windows and other operating systems, this can cause URLs to parse perfectly fine on Windows, but fail to parse on Linux. This function tries to parse the URL as a path on all operating systems.